Created A Habit Tracker.py
Thought I'd make a Habit Tracker thats simple but most importantly, FUN. Anyone seeing this small project of mine, join in on the coding to make improvements, I welcome those who want to make tech for the reason of making it fun so that we want to use it for ourselves EVERYDAY. Tech that matters to us. To that end, Here's my first entry, my first Project. Cheers :)